requirements.txt
Django==3.0.2
Cython
amqp
aniso8601
arrow
asgiref
async-timeout
atomicwrites
attrs
autobahn
Automat
Babel
backcall
billiard
celery
channels
channels-redis
chardet
constantly
coverage
daphne
decorator
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-extensions
django-filter
django-mathfilters
django-phonenumber-field
django-split-settings
djangorestframework
factory-boy
Faker
gunicorn
hyperlink
idna
incremental
ipython
ipython-genutils
jedi
kombu
Markdown
more-itertools
numpy
pandas
parso
pexpect
pickleshare
Pillow
pluggy
promise
prompt-toolkit
psycopg2
ptyprocess
py-moneyed
Pycco
Pygments
PyHamcrest
pystache
pytest
python-dateutil
pytz
singledispatch
six
smartypants
sqlparse
tblib
tex
traitlets
Twisted
txaio
uuid
vine
wcwidth
Werkzeug
zope.interface
